+++ lib/unparser/writer/send.rb
@@ -91,7 +91,11 @@ module Unparser
       end
 
       def parses_as_constant?
-        n_const?(Unparser.parse(selector.to_s))
+        test = Unparser.parse_either(selector.to_s).from_right do
+          fail InvalidNodeError.new("Invalid selector for send node: #{selector.inspect}", node)
+        end
+
+        n_const?(test)
       end
 
       def details